Refactor extract version-specific serializers
