+++ spec/requests/candidate_api/get_candidates_v1_2_spec.rb
@@ -16,6 +16,7 @@ RSpec.describe 'GET /candidate-api/v1.2/candidates', type: :request do
       :completed_application_form,
       2,
       candidate: candidate,
+      application_choices_count: 3,
     )
 
     get_api_request "#{root_path}?updated_since=#{CGI.escape(1.day.ago.iso8601)}", token: candidate_api_token
@@ -29,11 +30,13 @@ RSpec.describe 'GET /candidate-api/v1.2/candidates', type: :request do
     expect(response_data.first['application_status']).to eq(ProcessState.new(application_forms.first).state.to_s)
     expect(response_data.first['recruitment_cycle_year']).to eq(application_forms.first.recruitment_cycle_year)
     expect(response_data.first['submitted_at']).to eq(application_forms.first.submitted_at.iso8601)
+    expect(response_data.first['application_choices'].count).to eq(3)
 
     expect(response_data.second['id']).to eq(application_forms.second.id)
     expect(response_data.second['application_phase']).to eq(application_forms.second.phase)
     expect(response_data.second['application_status']).to eq(ProcessState.new(application_forms.second).state.to_s)
     expect(response_data.second['recruitment_cycle_year']).to eq(application_forms.second.recruitment_cycle_year)
     expect(response_data.second['submitted_at']).to eq(application_forms.second.submitted_at.iso8601)
+    expect(response_data.second['application_choices'].count).to eq(3)
   end
 end
